requirements.txt
# requirements.txt

boto3==1.26.90
requests==2.28.2
python-dotenv==1.0.0
colorama==0.4.6
termcolor==2.2.0
pyfiglet==0.8.post1